The Jupyter document file & input SDSS spectra
There are the original Jupyter document, input SDSS spectra, output *.cxt file (maybe you need to copy the output from print window, and the paste it into the *.cxt file. I recommend notepad++ for editing the *.cxt file), output spectra flux figures.
This is still a humble way of achieving the creation of the input file for starlight, and maybe "half-manual" code.